requirements.txt
Flask==3.0.0
flask-cors
flask-httpauth

pandas==2.2.3
joblib
scikit-learn==1.6.1

pydantic==2.11.4

matplotlib==3.10.1
cianparser==1.0.4


black==24.10.0
dvc==3.59.2


boto3
python-dotenv
flask-cors
boto3
flask_httpauth 
flask
flask-cors
flask-httpauth
gunicorn
boto3
python-dotenv
pydantic>=2.0
joblib
pandas
scikit-learn==1.6.1
